+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/util/JournalParams.java
@@ -44,6 +44,14 @@ public class JournalParams {
           JournalRecord.ActionStatus.COMPLETED));
       }
     },
+    DI_SRS_MARC_AUTHORITY_RECORD_UPDATED {
+      @Override
+      public Optional<JournalParams> getJournalParams(DataImportEventPayload eventPayload) {
+        return Optional.of(new JournalParams(JournalRecord.ActionType.UPDATE,
+          JournalRecord.EntityType.MARC_AUTHORITY,
+          JournalRecord.ActionStatus.COMPLETED));
+      }
+    },
     DI_SRS_MARC_BIB_RECORD_MODIFIED {
       @Override
       public Optional<JournalParams> getJournalParams(DataImportEventPayload eventPayload) {
@@ -65,6 +73,14 @@ public class JournalParams {
       public Optional<JournalParams> getJournalParams(DataImportEventPayload eventPayload) {
         return Optional.of(new JournalParams(JournalRecord.ActionType.NON_MATCH,
           JournalRecord.EntityType.MARC_BIBLIOGRAPHIC,
+          JournalRecord.ActionStatus.COMPLETED));
+      }
+    },
+    DI_SRS_MARC_AUTHORITY_RECORD_NOT_MATCHED {
+      @Override
+      public Optional<JournalParams> getJournalParams(DataImportEventPayload eventPayload) {
+        return Optional.of(new JournalParams(JournalRecord.ActionType.NON_MATCH,
+          JournalRecord.EntityType.MARC_AUTHORITY,
           JournalRecord.ActionStatus.COMPLETED));
       }
     },